# openapi.yaml
openapi: 3.0.3
info:
  title: BOWAS API
  version: 1.0.0
  description: REST API for WhatsApp messaging with JWT authentication.
servers:
- url: http://localhost:3000
  description: Default local server
paths:
  /health:
    get:
      operationId: getHealth
      summary: Health check
      responses:
        '200':
          description: API status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              examples:
                ready:
                  value:
                    status: ok
                    whatsappReady: true
                notReady:
                  value:
                    status: ok
                    whatsappReady: false
  /login:
    post:
      operationId: login
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              defaultCredentials:
                value:
                  username: admin
                  password: admin
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                success:
                  value:
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJhZG1pbiJ9.signature
                    tokenType: Bearer
                    expiresIn: 3600
        '400':
          description: Invalid JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidJson:
                  value:
                    error: Invalid JSON body
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCredentials:
                  value:
                    error: Invalid username or password
  /send:
    post:
      operationId: sendMessage
      summary: Send WhatsApp message
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendRequest'
            examples:
              numberFormat:
                value:
                  to: '6281234567890'
                  message: hello from API
              chatIdFormat:
                value:
                  to: 6281234567890@c.us
                  message: hello explicit chat id
      responses:
        '200':
          description: Message sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
              examples:
                success:
                  value:
                    ok: true
                    to: 6281234567890@c.us
                    id: ABGGFlA5FpafAgo6EhQ2M2Iw
        '400':
          description: Validation or JSON parse error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingFields:
                  value:
                    error: 'Required fields: to, message'
                invalidJson:
                  value:
                    error: Invalid JSON body
        '401':
          description: Missing or invalid bearer token
          headers:
            WWW-Authenticate:
              schema:
                type: string
              example: Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingBearer:
                  value:
                    error: 'Unauthorized: Missing Bearer token'
                invalidSignature:
                  value:
                    error: 'Unauthorized: Invalid signature'
        '503':
          description: WhatsApp client not ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notReady:
                  value:
                    error: WhatsApp client is not ready yet
        '500':
          description: Failed to send message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                sendFailure:
                  value:
                    error: Failed to send message
  /whatsapp/deauth:
    post:
      operationId: whatsappDeauth
      summary: Deauthenticate WhatsApp client
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Client deauthenticated and reinit requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeauthResponse'
              examples:
                success:
                  value:
                    ok: true
                    message: WhatsApp client deauthenticated. A new QR will be generated
                      shortly.
        '401':
          description: Missing or invalid bearer token
          headers:
            WWW-Authenticate:
              schema:
                type: string
              example: Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingBearer:
                  value:
                    error: 'Unauthorized: Missing Bearer token'
        '500':
          description: Failed to deauthenticate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                failure:
                  value:
                    error: Failed to deauthenticate client
  /whatsapp/auth:
    post:
      operationId: whatsappAuthQr
      summary: Get WhatsApp auth QR
      security:
      - bearerAuth: []
      parameters:
      - in: query
        name: text
        required: false
        schema:
          type: boolean
          default: false
        description: Set `true` to return ASCII QR text in JSON response.
      - in: query
        name: html
        required: false
        schema:
          type: boolean
          default: false
        description: Set `true` to return an HTML page containing QR image.
      responses:
        '200':
          description: QR returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthQrResponse'
              examples:
                image:
                  value:
                    mode: image
                    generatedAt: '2026-02-25T00:00:00.000Z'
                    qrImageDataUrl: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...
                text:
                  value:
                    mode: text
                    qr: "\u2588\u2588\u2588 ASCII QR \u2588\u2588\u2588"
                    generatedAt: '2026-02-25T00:00:00.000Z'
            text/html:
              schema:
                type: string
              examples:
                htmlPage:
                  value: '<!doctype html><html><body><img src="data:image/png;base64,..." /></body></html>'
        '401':
          description: Missing or invalid bearer token
          headers:
            WWW-Authenticate:
              schema:
                type: string
              example: Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingBearer:
                  value:
                    error: 'Unauthorized: Missing Bearer token'
        '400':
          description: Invalid query combination
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidCombination:
                  value:
                    error: HTML mode requires image QR. Use html=true without text=true
        '409':
          description: Already authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                alreadyAuth:
                  value:
                    error: WhatsApp client is already authenticated
        '503':
          description: QR not available yet
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notAvailable:
                  value:
                    error: QR is not available yet. Please retry in a few seconds
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthResponse:
      type: object
      required:
      - status
      - whatsappReady
      properties:
        status:
          type: string
          example: ok
        whatsappReady:
          type: boolean
          example: true
    LoginRequest:
      type: object
      required:
      - username
      - password
      properties:
        username:
          type: string
          example: admin
        password:
          type: string
          example: admin
    LoginResponse:
      type: object
      required:
      - token
      - tokenType
      - expiresIn
      properties:
        token:
          type: string
          description: JWT token (HS256)
        tokenType:
          type: string
          enum:
          - Bearer
          example: Bearer
        expiresIn:
          type: integer
          format: int32
          example: 3600
    SendRequest:
      type: object
      required:
      - to
      - message
      properties:
        to:
          type: string
          description: Phone number or WhatsApp chat id (@c.us / @g.us)
          example: '6281234567890'
        message:
          type: string
          minLength: 1
          example: hello from API
    SendResponse:
      type: object
      required:
      - ok
      - to
      - id
      properties:
        ok:
          type: boolean
          example: true
        to:
          type: string
          example: 6281234567890@c.us
        id:
          type: string
          nullable: true
          example: ABGGFlA5FpafAgo6EhQ2M2Iw
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
          example: Invalid JSON body
    DeauthResponse:
      type: object
      required:
      - ok
      - message
      properties:
        ok:
          type: boolean
          example: true
        message:
          type: string
          example: WhatsApp client deauthenticated. A new QR will be generated shortly.
    AuthQrResponse:
      type: object
      required:
      - mode
      - generatedAt
      properties:
        mode:
          type: string
          enum:
          - image
          - text
        qr:
          type: string
          nullable: true
        generatedAt:
          type: string
          format: date-time
        qrImageDataUrl:
          type: string
          nullable: true
          description: Data URL PNG QR image